% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ARE}
\alias{ARE}
\title{Absolute relative error}
\usage{
ARE(y, ypred)
}
\arguments{
\item{y}{vector}

\item{ypred}{vector}
}
\value{
numeric number. A measure of difference between y and ypred.
}
\description{
Absolute relative error
}
