% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceffEst.R
\name{ceffEst}
\alias{ceffEst}
\title{Temporal effect estimation within cluster}
\usage{
ceffEst(doy, rmat, cluster, doyeval = seq(min(doy), max(doy)), h.cov = 100,
  h.sigma2 = 300, weight.cov = NULL, weight.sigma2 = NULL,
  max.per.cluster = 30, pve = 0.99, t.grid = NULL, t.grid.num = 50)
}
\arguments{
\item{doy}{vector for day of year}

\item{rmat}{residual matrix with rows corresponding to `doy` and columns corresponding to pixel index}

\item{cluster}{non-negative integer numbers representing the cluster for each column of rmat belongs to.
0 represents mask.}

\item{doyeval}{a vector on which to estimate}

\item{h.cov}{bandwidth for temporal covariance estimation; ignored if `weight.cov` is supplied}

\item{h.sigma2}{bandwidth for sigma2 estimation}

\item{weight.cov}{weight vector for temporal covariance estimation}

\item{weight.sigma2}{weight vector for sigma2 estimation; not used for now...}

\item{max.per.cluster}{maximum number of pixels used tfor covariance estimation}

\item{pve}{percentage of variance explained; used for number of eigen values selection}

\item{t.grid}{vector on which to estimate covariance matrix}

\item{t.grid.num}{number of grid points on the interval of doyeval to use for covariance estimation. not used if t.grid is given.}
}
\description{
Temporal effect estimation within cluster
}
