% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{doyMeanEst}
\alias{doyMeanEst}
\title{DOY mean estimation}
\usage{
doyMeanEst(rst, doy.idx = 1:nlayers(rst))
}
\arguments{
\item{rst}{a *Raster object}

\item{doy.idx}{a vector of DOY index of the *Raster object}
}
\value{
a *Raster object
}
\description{
DOY mean estimation
}
