% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_landsat.R
\name{gapfill_landsat}
\alias{gapfill_landsat}
\title{Gapfilling for repeated measurements over years}
\usage{
gapfill_landsat(year, doy, mat, img.nrow, img.ncol, doyeval = 1:365,
  h.tcov = 100, h.tsigma2 = 300, h.scov = 2, h.ssigma2 = 2, nnr = 10,
  outlier.action = c("keep", "remove"), intermediate.save = TRUE,
  intermediate.dir = "./output/", teff = TRUE, seff = TRUE)
}
\arguments{
\item{year}{}

\item{doy}{}

\item{mat}{}

\item{img.nrow}{}

\item{img.ncol}{}

\item{doyeval}{}

\item{h.tcov}{}

\item{h.tsigma2}{}

\item{h.scov}{}

\item{h.ssigma2}{}

\item{nnr}{}

\item{outlier.action}{}

\item{intermediate.save}{}

\item{intermediate.dir}{}

\item{teff}{}

\item{seff}{}
}
\description{
Gapfilling for repeated measurements over years
}
