% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_modis.R
\name{gapfill_modis}
\alias{gapfill_modis}
\title{Title}
\usage{
gapfill_modis(doy, mat, img.nrow, img.ncol, doyeval = 1:365, h.tcov = 100,
  h.tsigma2 = 300, h.scov = 2, h.ssigma2 = 2, nnr = 10, msk = NULL,
  ncluster = 50, max.per.cluster = 30, t.grid.num = 50,
  breaks = list(block.nrow = 10, block.ncol = 10, img.nrow = 30, img.ncol =
  30), outlier.action = c("keep", "remove"), intermediate.save = TRUE,
  intermediate.dir = "./output/", use.intermediate.result = TRUE,
  teff = FALSE, seff = TRUE)
}
\arguments{
\item{doy}{}

\item{mat}{}

\item{img.nrow}{}

\item{img.ncol}{}

\item{doyeval}{}

\item{h.tcov}{}

\item{h.tsigma2}{}

\item{h.scov}{}

\item{h.ssigma2}{}

\item{nnr}{}

\item{msk}{vector of mask}

\item{ncluster}{number of clusters. If 0 no cluster analysis is performed and the mean is estimated based on
pixel level.}

\item{max.per.cluster}{maximum number of pixels to select for covariance estimation.}

\item{t.grid.num}{number of grid on doyeval, on which to estimate the covariance matrix.}

\item{breaks}{if NULL, estimate spatial effect on the whole image; if a list with 'block.nrow', 'block.ncol',
'img.nrow' and 'img.ncol', the image is broken down into small blocks and do estimation seperately.}

\item{outlier.action}{}

\item{intermediate.save}{TRUE or FALSE. Whether to save intermediate results.}

\item{intermediate.dir}{Path to save intermediate results.}

\item{use.intermediate.result}{whether to use the intermediate results in the 'intermediate.dir' folder instead
of recalcuating}

\item{teff}{TRUE or FALSE. Whether to estimate temporal effect.}

\item{seff}{TRUE or FALSE. Whether to estimate spatial effect.}
}
\description{
Title
}
