% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{getMask}
\alias{getMask}
\alias{getMask.matrix}
\alias{getMask.RasterStack}
\alias{getMask.RasterBrick}
\title{Get the missing pattern (mask) of x}
\usage{
getMask(x, ...)

\method{getMask}{matrix}(x, ...)

\method{getMask}{RasterStack}(x, ...)

\method{getMask}{RasterBrick}(x, ...)
}
\arguments{
\item{x}{a matrix, RasterStack or RasterBrick}
}
\value{
a vector or RasterLayer object with TRUE and FALSE, TRUE means missing
}
\description{
Get the missing pattern (mask) of x
}
