% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMissingLayers}
\alias{getMissingLayers}
\title{Get missing layer index}
\usage{
getMissingLayers(rst.list)
}
\arguments{
\item{rst.list}{a RasterStack or RasterBrick object or a list of them}
}
\value{
index of the missing layers
}
\description{
Get missing layer index
}
