% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsat106}
\alias{landsat106}
\alias{landsat2}
\title{Landsat data example}
\format{
A data frame with 990 rows and 963 columns:
\itemize{
  \item year year
  \item doy day of the year
  \item pixeli pixel value for the i-th pixel of the image
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 990 rows and 963 columns.
}
\usage{
landsat106

landsat2
}
\description{
A dataset containing observation values of a 31x31 pixcels 
landsat image observed between year 1982 and 2015.
}
\keyword{datasets}
