% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lc_cov_1d}
\alias{lc_cov_1d}
\title{Local constant covariance estimation}
\usage{
lc_cov_1d(ids, time, resid, W, t1, t2)
}
\arguments{
\item{ids}{a vector indicating subject/group ids}

\item{time}{integer vector of observed time points, the minimum time unit is 1}

\item{resid}{vector of residual values used for covariance calculation}

\item{W}{weight vector, it contains both kernel and bandwidth information in general 
local polynomial estimation setting up}

\item{t1}{time point 1}

\item{t2}{time point 2}
}
\description{
Local constant covariance estimation
}
