% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanEst.R
\name{llreg}
\alias{llreg}
\title{Local linear regression}
\usage{
llreg(x, y, x.eval = x, minimum.num.obs = 4, h = 60, Kern = epan)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{dnew data to predict on}

\item{minimum.num.obs}{minimum number of observations needed to run the regression}

\item{h}{bandwidth}

\item{Kern}{Kernel}
}
\value{
predicted values at 'x.eval'
}
\description{
Local linear regression
}
