% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCurve.R
\name{lpreg}
\alias{lpreg}
\title{Local Polynomial Regression}
\usage{
lpreg(x, y, x.eval, span, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{vector to predict on}

\item{span}{see 'loess' function}

\item{...}{other parameters passed to 'loess' function}
}
\value{
predicted values at 'x.eval'
}
\description{
Local Polynomial Regression
}
