% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mat2stack}
\alias{mat2stack}
\title{Matrix to stack object}
\usage{
mat2stack(mat, nrow, idx = 1:nrow(mat), byrow = FALSE, ...)
}
\arguments{
\item{mat}{each row corresponse to an image}

\item{nrow}{number of rows of the image}

\item{idx}{the row indexes of the mat that will be used to make the stack object}

\item{byrow}{passed to matrix function}

\item{...}{other parameters passed to 'raster' function}
}
\value{
a stack object
}
\description{
Matrix to stack object
}
