% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCurve.R
\name{meanCurve}
\alias{meanCurve}
\title{Mean curve by pixel}
\usage{
meanCurve(x, y, x.eval, plot = FALSE)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{x.eval}{vector to predict on}

\item{plot}{logical. NOT IMPLEMENTED YET}
}
\description{
Mean curve by pixel
}
