% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanEst.R
\name{meanEst}
\alias{meanEst}
\title{Local constant estimation for image mean}
\usage{
meanEst(mat, wmat)
}
\arguments{
\item{mat}{a matrix, each column is a stacked values of an image}

\item{wmat}{weight matrix}
}
\description{
Local constant estimation for image mean
}
