% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NMSE}
\alias{NMSE}
\title{Normalized Mean Square Estimation}
\usage{
NMSE(y, ypred)
}
\arguments{
\item{y}{vector}

\item{ypred}{vector}
}
\value{
numeric number. A measure of difference between y and ypred.
}
\description{
Normalized Mean Square Estimation
}
