% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\docType{data}
\name{opts}
\alias{opts}
\title{Options for gapfill}
\format{An object of class \code{list} of length 2.}
\usage{
opts
}
\description{
Options for gapfill
}
\examples{
opts$get('meanCurve')
}
\keyword{datasets}
