% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACE.R
\name{PACE}
\alias{PACE}
\title{The following function deals with the imputation using sparse pca.}
\usage{
PACE(mat, ev.vec, sigma2, ev.val)
}
\arguments{
\item{mat}{data matrix. Each row is a row stacked image}

\item{ev.vec}{eigen vectors}

\item{sigma2}{measurement error}

\item{ev.val}{eigen values}
}
\value{
an imputed mat matrix
}
\description{
The following function deals with the imputation using sparse pca.
}
