% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pctMissing}
\alias{pctMissing}
\title{Missing value percentages}
\usage{
pctMissing(x, mc.cores)
}
\arguments{
\item{x}{A \code{RasterStack} object}

\item{mc.cores}{Numer of cores to use}
}
\value{
A vector of percent of missing values for each layer
}
\description{
Missing value percentages
}
