% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{phi.interp}
\alias{phi.interp}
\title{eigen function interpolation}
\usage{
phi.interp(tt, phi.fun, t.grid)
}
\arguments{
\item{tt}{time vector to interpolate on}

\item{phi.fun}{eigen function}

\item{t.grid}{time vector on which the eigen function was calculated}
}
\value{
interpolated eigen function with the number of rows equal length of tt.
}
\description{
eigen function interpolation
}
