% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{phiInterp}
\alias{phiInterp}
\title{Eigen function interpolation}
\usage{
phiInterp(tt, phi.fun, t.grid)
}
\arguments{
\item{tt}{time vector to interpolate on}

\item{phi.fun}{eigen function}

\item{t.grid}{time vector on which the eigen function was calculated}
}
\value{
interpolated eigen function with the number of rows equal length of tt.
}
\description{
Eigen function interpolation
}
