% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{RMSE}
\alias{RMSE}
\alias{NMSE}
\alias{ARE}
\title{Error metrics}
\usage{
RMSE(y, ypred)

NMSE(y, ypred)

ARE(y, ypred)
}
\arguments{
\item{y}{vector}

\item{ypred}{vecotr}
}
\value{
numeric number. A measure of difference between y and ypred.
}
\description{
Error metrics
}
