% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seffEst.R
\name{seffEst}
\alias{seffEst}
\title{Gapfill missing values for a sequence of images}
\usage{
seffEst(rmat, img.nrow, img.ncol, h.cov = 2, h.sigma2 = 2,
  weight.cov = NULL, weight.sigma2 = NULL, nnr, method = c("lc", "emp"),
  keep.original = FALSE, pve = 0.99, msk = NULL, msk.tol = 0.95,
  detail = FALSE)
}
\arguments{
\item{rmat}{residual matrix}

\item{img.nrow}{image row dimension}

\item{img.ncol}{image column dimension}

\item{h.cov}{bandwidth for temporal covariance estimation; ignored if `weight.cov` is supplied}

\item{h.sigma2}{bandwidth for sigma2 estimation}

\item{weight.cov}{weight vector for temporal covariance estimation}

\item{weight.sigma2}{weight vector for sigma2 estimation; not used for now...}

\item{nnr}{maximum of nearest neibors used to calculate correlation}

\item{method}{"lc" for local constant covariance estimation and "emp" for empirical covariance estimation}

\item{keep.original}{whether to keep the originally observed values in the resulting matrix}

\item{pve}{percent of variance explained of the selected eigen values}

\item{msk}{optional logistic matrix. TRUE for mask values.}

\item{msk.tol}{if 'msk' is not given, the program will determine the mask using 'getMask'
function. If the percentage of missing values for a pixel over time is greater than this}

\item{detail}{if TRUE returns eigen values,eigen functions}
}
\value{
a list containing the following components:
`year` same as input `year`
`doy` same as input `doy`
`imputed.partial`
}
\description{
Gapfill missing values for a sequence of images
}
