% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{similarity_mat}
\alias{similarity_mat}
\title{get similarity matrix}
\usage{
similarity_mat(mat.list)
}
\arguments{
\item{mat.list}{a list of matrix}
}
\value{
a matrix consists of the index of the matrices. For example, supose there are three matrices in the
list and the first column is c(3,2) which means matrix 3 is more similar to matrix 1 than matrix 2.
}
\description{
get similarity matrix
}
