% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCurve.R
\name{smooth_spline}
\alias{smooth_spline}
\title{Smoothing spline regression}
\usage{
smooth_spline(x, y, x.eval = x, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{vector to predict on}

\item{plot}{logical. NOT IMPLEMENTED YET}
}
\value{
predicted values at 'x.eval'
}
\description{
Smoothing spline regression
}
