% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCurve.R
\name{spreg}
\alias{spreg}
\title{spline regression}
\usage{
spreg(x, y, x.eval, minimum.num.obs = 4, basis = c("fourier", "bspline"),
  rangeval = c(min(x.eval) - 1, max(x.eval)), nbasis = 11, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{vector to predict on}

\item{nbasis}{see \code{fda::create.basis}}

\item{...}{arguments passed to \code{fad::create.basis} functions}

\item{rangeeval}{see \code{fda::create.basis}}
}
\value{
predicted values at 'x.eval'
}
\description{
spline regression
}
