% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfit-package.R
\docType{package}
\name{stfit-package}
\alias{stfit-package}
\title{stfit: Spatial-Temporal Functional Imputation Tool}
\description{
The stfit package provides functions to impute missing values for a 
sequence of observed images for the same location using functional
data analysis technique
}
