% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffEst.R
\name{teffEst}
\alias{teffEst}
\title{Temporal effect estimation}
\usage{
teffEst(
  ids,
  doy,
  rmat,
  doyeval = seq(min(doy), max(doy)),
  h.cov = 100,
  h.sigma2 = 300,
  weight.cov = NULL,
  weight.sigma2 = NULL,
  pve = 0.99,
  t.grid = NULL,
  t.grid.num = 50
)
}
\arguments{
\item{ids}{ids for 'group', for data with repeated measurement over years, year is ids; 
for pixels belong to certain clusters, cluster is ids.}

\item{doy}{vector for day of year}

\item{rmat}{residual matrix with rows corresponding to `doy` and columns corresponding to pixel index}

\item{doyeval}{a vector on which to estimate}

\item{h.cov}{bandwidth for temporal covariance estimation; ignored if `weight.cov` is supplied}

\item{h.sigma2}{bandwidth for sigma2 estimation}

\item{weight.cov}{weight vector for temporal covariance estimation}

\item{weight.sigma2}{weight vector for sigma2 estimation}

\item{pve}{percentage of variance explained; used for number of eigen values selection}

\item{t.grid}{vector of grid points on which to calculate the covariance function}

\item{t.grid.num}{number of grid points to use for covariance estimation. Ignored if `t.grid` is given.}
}
\value{

}
\description{
Temporal effect estimation
}
