% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{tod_impute}
\alias{tod_impute}
\title{Image imputation using time of the day information}
\usage{
tod_impute(rst.list, sim.mat)
}
\arguments{
\item{rst.list}{A list of `RasterBrick` or `RasterStack` objects, whose elements are
measurements of the same location at different time of the day of the same day of 
the year}
}
\description{
Image imputation using time of the day information
}
