% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightMatrix.R
\name{weightVector}
\alias{weightVector}
\title{weight vector calculation}
\usage{
weightVector(h)
}
\arguments{
\item{h}{bandwidth, should be positive numbers}
}
\value{
a vector
}
\description{
weight vector calculation
}
