% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill.R
\name{gapfill}
\alias{gapfill}
\title{Gapfill missing values for a sequence of images}
\usage{
gapfill(year, doy, mat, img.nrow, img.ncol, h, doyrange = seq(min(doy),
  max(doy)), nnr, method = c("lc", "emp"), pve = 0.99, outlier.tol = 0.2,
  outlier.action = c("ac", "keep"), mc.cores = parallel::detectCores())
}
\arguments{
\item{year}{vector of year index}

\item{doy}{vector of day of year (DOY) index}

\item{mat}{data matrix. Each row is a row stacked image}

\item{img.nrow}{image row dimension}

\item{img.ncol}{image column dimension}

\item{h}{bandwidth}

\item{doyrange}{vector of two elements, minimum and maximum value of DOY}

\item{pve}{percent of variance explained of the selected eigen values}

\item{mc.cores}{number of cores to use for computation}
}
\value{
a list containing the following components:
`year` same as input `year`
`doy` same as input `doy`
`imputed.partial`
}
\description{
Gapfill missing values for a sequence of images
}
