% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMask}
\alias{getMask}
\title{Get image mask}
\usage{
getMask(object, tol = 0.95)
}
\arguments{
\item{object}{A numeric matrix. Each row is an row stacked image.}

\item{tol}{If the percentage of missing values for a pixel over time is greater than this
value, this pixel is treated as a mask value.}
}
\description{
Get image mask
}
