% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMask}
\alias{getMask}
\alias{getMask,matrix-method}
\title{Get the missing pattern (mask) of x}
\usage{
getMask(object, tol)

\S4method{getMask}{matrix}(object, tol = 0.95)
}
\arguments{
\item{object}{each row is am image; each column is the stacked values of an image}

\item{tol}{If the percentage of missing values for a pixel over time is greater than this
value, this pixel is treated as a mask value.}
}
\description{
Get the missing pattern (mask) of x
}
