% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\docType{methods}
\name{getMask}
\alias{getMask}
\alias{getMask,matrix-method}
\title{Get the missing pattern (mask) of x}
\usage{
getMask(object, ...)

\S4method{getMask}{matrix}(object, tol = 0.95, ...)
}
\arguments{
\item{tol}{If the percentage of missing values for a pixel over time is greater than this
value, this pixel is treated as a mask value.}

\item{matrix}{each row is am image; each column is the stacked values of an image}
}
\description{
Get the missing pattern (mask) of x
}
