% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanEst.R
\name{lpreg}
\alias{lpreg}
\title{Local Polynomial Regression}
\usage{
lpreg(x, y, x.eval, minimum.num.obs = 4, span = 0.3, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{vector to predict on}

\item{minimum.num.obs}{minimum number of observations needed to run the regression}

\item{span}{see 'loess' function}

\item{...}{other parameters passed to 'loess' function}
}
\value{
predicted values at 'x.eval'
}
\description{
Local Polynomial Regression
}
