% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanEst.R
\name{meanEst}
\alias{meanEst}
\title{Gapfill missing values for a sequence of images}
\usage{
meanEst(doy, mat, doyeval = seq(min(doy), max(doy)), msk = rep(FALSE,
  ncol(mat)), outlier.tol = 0.5, minimum.num.obs = 4, cluster = NULL,
  redo = TRUE)
}
\arguments{
\item{doy}{vector of day of year (DOY) index}

\item{mat}{data matrix. Each row is a row stacked image}

\item{doyeval}{vector of doy to evaluate on}

\item{msk}{optional logistic matrix. TRUE for mask values.}

\item{outlier.tol}{the tolerance value in defining an image as outlier. The percent of 
outlier pixels in an image exceed this value is regarded as outlier image which will not
be used in temporal mean estimation.}

\item{minimum.num.obs}{minimum number of observations needed for doing mean estimation for each pixel}

\item{cluster}{optional matrix defining clusters of pixels. If NULL, temporal mean estimation
is conducted on each pixel, otherwise all pixels from the same cluster are combined for
temporal mean estimation.}
}
\value{
a list containing the following components:
`doyeval` same as input `doyeval`
`meanmat` matrix with number of rows equal length of `doyeval` and number of columns equal `ncol(mat)`
}
\description{
Gapfill missing values for a sequence of images
}
