% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanEst.R
\name{smooth_spline}
\alias{smooth_spline}
\title{Smoothing spline regression}
\usage{
smooth_spline(x, y, x.eval = x, minimum.num.obs = 4, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{response variable}

\item{x.eval}{vector to predict on}

\item{minimum.num.obs}{minimum number of observations needed to run the regression}

\item{...}{other parameters to be passed to \code{smooth.spline} function}
}
\value{
predicted values at 'x.eval'
}
\description{
Smoothing spline regression
}
