% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightMatrix.R
\name{weightMatrix}
\alias{weightMatrix}
\title{weight matrix calculation}
\usage{
weightMatrix(h)
}
\arguments{
\item{h}{'bandwith'}
}
\value{
a weighting matrix
}
\description{
weight matrix calculation
}
